+++ CHANGELOG.md
@@ -2,14 +2,12 @@
 
 ## 1.1.1 - 2018-09-24
 
-### Changed
-
+### Fixed
 - Fix issue where classes were being incorrectly applied in IE11.
 
 ## 1.1.0 - 2018-09-10
 
 ### Added
-
 - Added ability to pass in a country code to the `tel` filter to parse national numbers.
 
 ## 1.0.0 - 2018-09-05