+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 1.2.0 - 2018-10-19
 
 ### Added
 
@@ -10,11 +10,13 @@
 ## 1.1.1 - 2018-09-24
 
 ### Fixed
+
 - Fix issue where classes were being incorrectly applied in IE11.
 
 ## 1.1.0 - 2018-09-10
 
 ### Added
+
 - Added ability to pass in a country code to the `tel` filter to parse national numbers.
 
 ## 1.0.0 - 2018-09-05